pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ytdlp-gui"
version = "1.0.0"
description = "A professional desktop GUI wrapper for yt-dlp"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "you@example.com"},
]
keywords = ["yt-dlp", "youtube", "downloader", "gui", "desktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "yt-dlp>=2025.1.0",
    "customtkinter>=5.2.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/yt-dlp-gui"
Repository = "https://github.com/yourusername/yt-dlp-gui"
"Bug Tracker" = "https://github.com/yourusername/yt-dlp-gui/issues"

[project.scripts]
ytdlp-gui = "ytdlp_gui.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
